// buddy-script-backend/src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PostPrivacy = {
  public: 'public',
  private: 'private'
} as const

export type PostPrivacy = (typeof PostPrivacy)[keyof typeof PostPrivacy]


export const LikeTargetType = {
  post: 'post',
  comment: 'comment',
  reply: 'reply'
} as const

export type LikeTargetType = (typeof LikeTargetType)[keyof typeof LikeTargetType]